+++ oauth2/doc.go
@@ -1,0 +1,32 @@
+//      ┌─────────┐                ┌───────────┐                     ┌────────┐
+//      │ Browser │                │Chronograf │                     │Provider│
+//      └─────────┘                └───────────┘                     └────────┘
+//           │                           │                                │
+//           ├─────── GET /auth ─────────▶                                │
+//           │                           │                                │
+//           │                           │                                │
+//           ◀ ─ ─ ─302 to Provider  ─ ─ ┤                                │
+//           │                           │                                │
+//           │                           │                                │
+//           ├──────────────── GET /auth w/ callback ─────────────────────▶
+//           │                           │                                │
+//           │                           │                                │
+//           ◀─ ─ ─ ─ ─ ─ ─   302 to Chronograf Callback  ─ ─ ─ ─ ─ ─ ─ ─ ┤
+//           │                           │                                │
+//           │   Code and State from     │                                │
+//           │        Provider           │                                │
+//           ├───────────────────────────▶    Request token w/ code &     │
+//           │                           │             state              │
+//           │                           ├────────────────────────────────▶
+//           │                           │                                │
+//           │                           │          Response with         │
+//           │                           │              Token             │
+//           │   Set cookie, Redirect    │◀ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ┤
+//           │           to /            │                                │
+//           ◀───────────────────────────┤                                │
+//           │                           │                                │
+//           │                           │                                │
+//           │                           │                                │
+//           │                           │                                │
+
+package oauth2
